Update index.ts
I have added 3 functions to add value to this canister, check them out.

## Functions

### `updateSaving(id: string, updatedPayload: SavingPayload): Result<Saving, string>`

Updates an existing saving record with new information. Users can provide the ID of the saving record they want to update and the updated payload containing the new values. The function returns a Result containing the updated saving record if the update is successful or an error message if the record is not found.

### `calculateInterest(days: number, amount: number, specifiedYear: number): Result<number, string>`

Calculates the interest earned on a saving based on the provided parameters. Users can specify the number of days the saving is held, the amount of the saving, and the specified year for the calculation. The function returns a Result containing the calculated interest or an error message.

### `transferSaving(id: string, recipient: string): Result<Saving, string>`

Transfers a saving record to another user. Users can provide the ID of the saving record they want to transfer and the username of the recipient. The function updates the ownership of the saving record and transfers it to the specified recipient. It returns a Result containing the updated saving record if the transfer is successful or an error message if the record is not found.